requirements.txt
aiogram>=3.0.0
sqlalchemy>=2.0.0
alembic>=1.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
python-dotenv>=1.0.0
environs>=9.0.0
aiosqlite>=0.19.0 